Save issue fixed, improved file handler
Save issue has been fixed, it would fail if you hadn't opened a fie before. This has now been fixed with a simple test, so if haven't it just brings up save as

The file handler has been improved, the charset is now set within the handler (instead of in the function for open and save), and maybe in future added as an option in the constructor. The file chooser stuff is now handed off the file handler to deal with. And more check have been put in place to do with file handling in general. The handler it self should now be easier to deal with, hopefully at least